+++ generators/custombot/index.js
@@ -64,7 +64,7 @@ module.exports =
 /******/ 	__webpack_require__.p = "";
 /******/
 /******/ 	// Load entry module and return exports
-/******/ 	return __webpack_require__(__webpack_require__.s = 18);
+/******/ 	return __webpack_require__(__webpack_require__.s = 19);
 /******/ })
 /************************************************************************/
 /******/ ([
@@ -142,40 +142,24 @@ module.exports = require("yeoman-generator");
 module.exports = require("yosay");
 
 /***/ }),
-/* 6 */
-/***/ (function(module, exports, __webpack_require__) {
-
-"use strict";
-
-Object.defineProperty(exports, "__esModule", { value: true });
-/**
- * Configuration options for the generator
- */
-class GeneratorTeamTabOptions {
-    constructor() {
-        this.botType = "";
-    }
-}
-exports.GeneratorTeamTabOptions = GeneratorTeamTabOptions;
-
-
-/***/ }),
+/* 6 */,
 /* 7 */,
-/* 8 */
+/* 8 */,
+/* 9 */
 /***/ (function(module, exports, __webpack_require__) {
 
 "use strict";
 
 Object.defineProperty(exports, "__esModule", { value: true });
-const BotGenerator_1 = __webpack_require__(12);
-module.exports = BotGenerator_1.BotGenerator;
+const CustomBotGenerator_1 = __webpack_require__(13);
+module.exports = CustomBotGenerator_1.CustomBotGenerator;
 
 
 /***/ }),
-/* 9 */,
 /* 10 */,
 /* 11 */,
-/* 12 */
+/* 12 */,
+/* 13 */
 /***/ (function(module, exports, __webpack_require__) {
 
 "use strict";
@@ -183,146 +167,58 @@ module.exports = BotGenerator_1.BotGenerator;
 Object.defineProperty(exports, "__esModule", { value: true });
 const Generator = __webpack_require__(4);
 const lodash = __webpack_require__(3);
-const GeneratorTeamTabOptions_1 = __webpack_require__(6);
 const Yotilities_1 = __webpack_require__(1);
 let yosay = __webpack_require__(5);
 let path = __webpack_require__(0);
 let Guid = __webpack_require__(2);
-class BotGenerator extends Generator {
+class CustomBotGenerator extends Generator {
     constructor(args, opts) {
         super(args, opts);
         opts.force = true;
-        this.options = opts.options === undefined ? new GeneratorTeamTabOptions_1.GeneratorTeamTabOptions() : opts.options;
-        this.desc('Adds a Bot to a Teams project.');
+        this.options = opts.options;
+        this.desc('Adds a custom bot to a Teams project.');
     }
     prompting() {
-        if (this.options.bot) {
+        if (this.options.customBot) {
             return this.prompt([
                 {
-                    type: 'list',
-                    name: 'bottype',
-                    message: 'Would type of bot would you like to use?',
-                    default: 'existing',
-                    choices: [
-                        {
-                            name: 'An already existing and running bot',
-                            value: 'existing'
-                        },
-                        {
-                            name: 'A new Bot Framework bot',
-                            value: 'botframework'
-                        }
-                    ]
-                },
-                {
                     type: 'input',
-                    name: 'botname',
-                    message: 'What is the name of your bot?',
-                    default: this.options.title + 'bot',
-                    validate: (input) => {
-                        return input.length > 0;
-                    },
-                    when: (answers) => answers.bottype != 'existing'
+                    name: 'title',
+                    message: 'Name of your custom bot?',
+                    default: this.options.title + 'custombot'
                 },
-                {
-                    type: 'input',
-                    name: 'botid',
-                    message: (answers) => {
-                        var message = 'I need an ID for the bot. ';
-                        if (answers.botTye == 'botframework') {
-                            message += 'If you don\'t specify a value now, you need to manually edit it later. ';
-                        }
-                        message += 'It\'s found in the Bot Framework portal';
-                        return message;
-                    },
-                    default: (answers) => {
-                        if (answers.bottype == 'botframework') {
-                            return Guid.EMPTY;
-                        }
-                        return '';
-                    },
-                    validate: (input) => {
-                        return Guid.isGuid(input);
-                    }
-                },
-                {
-                    type: 'confirm',
-                    name: 'pinnedTab',
-                    message: 'Do you want to add a pinned tab to your bot?',
-                },
-                {
-                    type: 'input',
-                    name: 'pinnedTabName',
-                    message: 'What is the title of your pinned tab for the bot?',
-                    when: (answers) => {
-                        return answers.pinnedTab;
-                    }
-                }
             ]).then((answers) => {
-                this.options.botid = answers.botid;
-                this.options.pinnedTab = answers.pinnedTab;
-                this.options.pinnedTabTitle = answers.pinnedTabName;
-                this.options.pinnedTabName = lodash.camelCase(answers.pinnedTabName);
-                this.options.botType = answers.bottype;
-                this.options.botTitle = answers.botname;
-                this.options.botName = lodash.camelCase(answers.botname);
+                this.options.customBotTitle = answers.title;
+                this.options.customBotName = lodash.camelCase(answers.title);
             });
         }
     }
     writing() {
-        if (this.options.bot) {
-            let manifestPath = "src/manifest/manifest.json";
-            var manifest = this.fs.readJSON(manifestPath);
-            var newbot = {
-                mri: this.options.botid,
-                pinnedTabs: []
-            };
+        if (this.options.customBot) {
+            let templateFiles = [
+                "README-{customBotName}.md",
+                "src/app/{customBotName}.ts"
+            ];
             this.sourceRoot();
-            let templateFiles = [];
-            if (this.options.pinnedTab) {
-                templateFiles.push("src/app/scripts/{pinnedTabName}Tab.ts", "src/app/web/{pinnedTabName}Tab.html");
-                newbot.pinnedTabs.push({
-                    id: Guid.raw(),
-                    definitionId: Guid.raw(),
-                    displayName: this.options.pinnedTabTitle,
-                    url: `${this.options.host}/${this.options.pinnedTabName}Tab.html`,
-                    website: `${this.options.host}/${this.options.pinnedTabName}Tab.html`,
-                });
-            }
-            manifest.bots.push(newbot);
-            this.fs.writeJSON(manifestPath, manifest);
-            if (this.options.botType != 'existing') {
-                templateFiles.push('src/app/{botName}.ts');
-                templateFiles.push('README-{botName}.md');
-            }
             templateFiles.forEach(t => {
                 this.fs.copyTpl(this.templatePath(t), Yotilities_1.Yotilities.fixFileNames(t, this.options), this.options);
             });
-            // update client.ts
-            if (this.options.pinnedTab) {
-                let clientTsPath = "src/app/scripts/client.ts";
-                let clientTs = this.fs.read(clientTsPath);
-                clientTs += `\n// Added by generator-teams`;
-                clientTs += `\nexport * from './${this.options.pinnedTabName}Tab';`;
-                clientTs += `\n`;
-                this.fs.write(clientTsPath, clientTs);
-            }
         }
     }
 }
-exports.BotGenerator = BotGenerator;
+exports.CustomBotGenerator = CustomBotGenerator;
 
 
 /***/ }),
-/* 13 */,
 /* 14 */,
 /* 15 */,
 /* 16 */,
 /* 17 */,
-/* 18 */
+/* 18 */,
+/* 19 */
 /***/ (function(module, exports, __webpack_require__) {
 
-module.exports = __webpack_require__(8);
+module.exports = __webpack_require__(9);
 
 
 /***/ })