+++ src/app/GeneratorTeamTabOptions.ts
@@ -1,3 +1,6 @@
+/**
+ * Configuration options for the generator
+ */
 export class GeneratorTeamTabOptions {
     id: string;
     title: string;
@@ -6,8 +9,6 @@ export class GeneratorTeamTabOptions {
     name: string;
     shouldUseSubDir: boolean;
     libraryName: string;
-    shouldUseAzure: boolean;
-    shouldUseExpress: boolean;
     namespace: string;
     developer: string;
     privacy: string;
@@ -21,4 +22,11 @@ export class GeneratorTeamTabOptions {
     pinnedTab: boolean;
     pinnedTabName: string;
     pinnedTabTitle: string;
+    botType: string = "";
+    botTitle: string;
+    botName: string;
+    /* Custom Bot */
+    customBot: boolean;
+    customBotTitle: string;
+    customBotName: string;
 }