+++ src/app/templates/gulpfile.js
@@ -5,6 +5,7 @@ var gutil = require('gulp-util');
 var inject = require('gulp-inject');
 var runSequence = require('run-sequence');
 const zip = require('gulp-zip');
+var nodemon = require('nodemon');
 
 var injectSources = ["./dist/web/scripts/**/*.js"]
 var typeScriptFiles = ["./src/**/*.ts"]
@@ -14,16 +15,26 @@ var watcherfiles = ["./src/**/*.*"]
 var manifestFiles = ["./src/manifest/**/*.*"]
 
 
+/**
+ * Watches source files and invokes the build task
+ */
 gulp.task('watch', function () {
     gulp.watch('./src/**/*.*', ['build']);
 });
 
+
+/**
+ * Creates the tab manifest
+ */
 gulp.task('manifest', () => {
     gulp.src(manifestFiles)
         .pipe(zip('<%=solutionName%>.zip'))
         .pipe(gulp.dest('package'))
 });
 
+/**
+ * Webpack bundling
+ */
 gulp.task('webpack', function (callback) {
     var webpackConfig = require(process.cwd() + '/webpack.config')
     webpack(webpackConfig
@@ -50,17 +61,22 @@ gulp.task('webpack', function (callback) {
         });
 });
 
+/**
+ * Copies static files
+ */
 gulp.task('static:copy', function () {
     return gulp.src(staticFiles, { base: "./src/app" })
         .pipe(gulp.dest('./dist/'));
 })
 
+/**
+ * Injects script into pages
+ */
 gulp.task('static:inject', ['static:copy'], function () {
 
     var injectSrc = gulp.src(injectSources);
 
     var injectOptions = {
-        ignorePath: '/bin',
         relative: false,
         ignorePath: 'dist/web',
         addRootSlash: false
@@ -71,7 +87,25 @@ gulp.task('static:inject', ['static:copy'], function () {
         .pipe(gulp.dest('./dist'));
 });
 
-
+/**
+ * Build task, that uses webpack and injects scripts into pages
+ */
 gulp.task('build', function () {
     runSequence('webpack', 'static:inject')
+});
+
+/**
+ * Task for local debugging
+ */
+gulp.task('serve', ['build', 'watch'], function (cb) {
+    var started = false;
+    return nodemon({
+        script: 'dist/server.js',
+        watch: ['dist/server.js']
+    }).on('start', function () {
+        if (!started) {
+            cb();
+            started = true;
+        }
+    });
 });