+++ src/custombot/index.ts
@@ -1,0 +1,3 @@
+import {CustomBotGenerator} from './CustomBotGenerator';
+
+module.exports = CustomBotGenerator